+++ HISTORY.rst
@@ -5,15 +5,15 @@ History
 TODO
 ^^^^
 
-* Move imported binary file in ``interim`` subfolder
+* Add features to samples (relying on metadata file)
 * make rule to clean-up ``interim`` data
-* skip already processed file from import
 * Breed name should be indexed as text
 * Breed name should be a unique key within species
 
 0.2.2.dev0
 ----------
 
+* skip already processed file from import
 * Deal with ``mother_id`` and ``father_id`` (search for ``smarter_id`` in database)
 * Deal with multi-countries dataset
   - track country in aliases while importing breeds from dataset